requirements.txt
asyncio
discord.py
python-dotenv
mcstatus